Project 10 - Add AST Parser
Modify the Compiler module to parse a token stream lexed from a file
into a Jack Abstract Syntax Tree, with classes as the root nodes.

While this is a complete parser, we are back to writing Haskell types to
stdout instead of XML to files alongside the original Jack files. We
will either write a renderer for the AST which generates the XML, or we
could either modify the Parser to wrap our XML monad or keep the
generated XML tags in the Parser state.